refactor: reorganize tests to align with repo conventions
